randomCleanUp - DC-798 - DC-863 - DC-864 - DC-879 - DC-846 - DC-851 - DC-828

+++ DDI/DDI.UI.Web/Scripts/data.js
@@ -369,8 +369,9 @@ function LoadGridWithData(grid, container, columns, route, selected, editMethod,
                     .attr('title', 'Delete')
                     .click(function (e) {
                         e.preventDefault();
-
-                        deleteMethod(options.data.Id);
+                        ConfirmModal('Are you sure you want to delete this item?', function () {
+                            deleteMethod(options.data.Id);
+                        }, null); 
                     })
                     .appendTo(container);
             }
